Dental/streamlit/requirements.txt
streamlit

torch==2.3.1
torchvision==0.18.1

numpy==1.26.4
pandas==2.2.2
pillow==10.4.0

opencv-python-headless==4.10.0.84
scikit-learn==1.5.1
matplotlib==3.9.0
